Fix Note interface's deserialization.
Created a custom deserializer for Note to be deserialized into NoteImp
objects. For that to be possible, all references of Book inside of
NoteImp were removed. Note verification (page number) was also moved to
BookImp.

NoteDeserializer was used in Create, and Update where needed.

Also updated Book's and Note's tests to construct correct Note objects.