Create TestUtil + minor test related fixes.
TestUtil contains utility methods to be used for testing, and fixed
CreateTest to work with recent updates.

Also updated ClientManager to reinstantiate a new manager object every
time a setup is done to avoid infinite blocking (due to the manager
thread being dead). This was not necessary when manager ran only once,
but is necessary for testing as setup and clean are called several
times.