Server: Fix createBook to handle existance of labels.
createBook failed to add books to user's table if the user had labels.

This was caused by the fact that the number of columns in the used
INSERT statement doesn't match those in user's table, due to the
addition of new columns representing the added labels.

This is solved by specifying the names of the columns to insert. This
doesn't affect the values of the label columns, as they default to
false.